refact : Member 인수테스트 Request분리
